Extract AddEditEmailSidebar into file
