+++ frontend/src/metabase/sharing/components/AddEditSidebar/Heading.jsx
@@ -1,0 +1,8 @@
+import React from "react";
+import PropTypes from "prop-types";
+
+const Heading = ({ children }) => <h4>{children}</h4>;
+
+Heading.propTypes = { children: PropTypes.any };
+
+export default Heading;
