+++ frontend/src/metabase/sharing/components/AddEditSidebar/constants.js
@@ -1,0 +1,6 @@
+import { t } from "ttag";
+
+export const CHANNEL_NOUN_PLURAL = {
+  email: t`Emails`,
+  slack: t`Slack messages`,
+};
