+++ frontend/src/metabase/sharing/components/AddEditSidebar/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./AddEditSidebar";
